python/trashcord/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
discord.py==1.5.0
duckpy==3.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
httpcore==0.10.2
httpx==0.14.3
hyperframe==5.2.0
idna==2.10
lxml==4.5.2
multidict==4.7.6
rfc3986==1.4.0
sniffio==1.1.0
soupsieve==2.0.1
yarl==1.6.0
